+++ resources/views/messages/_widget.blade.php
@@ -77,7 +77,7 @@
 
         {{-- THREAD --}}
         <div x-show="view === 'thread'" class="flex-1 flex flex-col min-h-0">
-            <div x-ref="scroll" class="flex-1 overflow-y-auto p-3 space-y-1.5 bg-gray-50 dark:bg-gray-900/30">
+            <div x-ref="scroll" class="flex-1 overflow-y-auto p-3 space-y-2.5 bg-gray-50 dark:bg-gray-900/30">
                 <template x-for="m in messages" :key="m.id">
                     <div class="flex flex-col" :class="m.mine ? 'items-end' : 'items-start'">
                         <div class="px-3 py-1.5 rounded-2xl text-sm leading-snug shadow-sm"